feat(synapse-interface): create template event maintenance components, remove ecotone + metis event maintenance components (#2294)
* Remove upgrade components from app

* Create Event example for annoucements + progress bar, add comments

* Add comments for example of hook

* Adjust comments

* Adjust comments

* annoucement banner visual tweaks

* Add comment to show where progress bar was last used

* reset ECOTONE_FORK_END_DATE

* Add additional example code snippets

* Reorder imports

* Improve comments

---------
