+++ packages/synapse-interface/components/Maintenance/AnnouncementBanner.tsx
@@ -54,28 +54,16 @@ export const AnnouncementBanner = ({
   if (!showBanner || !hasMounted || !isStarted || isComplete) return null
 
   return (
-    <div
-      className="flex items-center justify-center mx-auto mb-3 lg:flex-row"
-      style={{
-        background:
-          'linear-gradient(310.65deg, rgba(172, 143, 255, 0.2) -17.9%, rgba(255, 0, 255, 0.2) 86.48%)',
-      }}
-    >
+    <div className="flex items-center justify-center mx-auto lg:flex-row text-sm text-left bg-gradient-to-r from-fuchsia-600/25 to-purple-600/25">
       <div
         id="banner-default"
-        className="flex items-center px-8 pt-1 pb-1 rounded-md text-primaryTextColor"
+        className="flex items-center gap-4 px-4 py-2 w-full max-w-[1111px] text-primaryTextColor justify-between leading-normal"
         role="alert"
       >
-        <div className="container p-1 mx-auto">
-          <p className="text-md">{bannerContents}</p>
-        </div>
+        {bannerContents}
         <button
           type="button"
-          className={`
-            inline-flex items-center justify-center
-            h-7 w-7 p-1.5 ml-3
-            text-primaryTextColor
-          `}
+          className="inline-flex items-center justify-center text-primaryTextColor"
           data-dismiss-target="#banner-default"
           aria-label="Close"
           onClick={() => setShowBanner(false)}
