+++ packages/synapse-interface/pages/index.tsx
@@ -3,8 +3,7 @@ import { Portfolio } from '@/components/Portfolio/Portfolio'
 import { LandingPageWrapper } from '@/components/layouts/LandingPageWrapper'
 import ReactGA from 'react-ga'
 import useSyncQueryParamsWithBridgeState from '@/utils/hooks/useSyncQueryParamsWithBridgeState'
-import { EcotoneForkUpgradeBanner } from '@/components/Maintenance/Events/EcotoneForkUpgrade'
-import { MetisDowntimeBanner } from '@/components/Maintenance/Events/MetisUpgrade'
+import { EcotoneForkUpgradeBanner } from '@/components/Maintenance/Events/example/EcotoneForkUpgrade'
 
 // TODO: someone should add this to the .env, disable if blank, etc.
 // this is being added as a hotfix to assess user load on the synapse explorer api
@@ -16,19 +15,20 @@ const Home = () => {
   useSyncQueryParamsWithBridgeState()
 
   return (
-    <LandingPageWrapper>
-      <main
-        data-test-id="bridge-page"
-        className="relative z-0 flex-1 h-full overflow-y-auto focus:outline-none"
-      >
-        <EcotoneForkUpgradeBanner />
-        <MetisDowntimeBanner />
-        <div className="flex flex-col-reverse justify-center gap-16 px-4 py-20 mx-auto lg:flex-row 2xl:w-3/4 sm:mt-6 sm:px-8 md:px-12">
-          <Portfolio />
-          <StateManagedBridge />
-        </div>
-      </main>
-    </LandingPageWrapper>
+    <>
+      <EcotoneForkUpgradeBanner />
+      <LandingPageWrapper>
+        <main
+          data-test-id="bridge-page"
+          className="relative z-0 flex-1 h-full overflow-y-auto focus:outline-none"
+        >
+          <div className="flex flex-col-reverse justify-center gap-16 px-4 py-20 mx-auto lg:flex-row 2xl:w-3/4 sm:mt-6 sm:px-8 md:px-12">
+            <Portfolio />
+            <StateManagedBridge />
+          </div>
+        </main>
+      </LandingPageWrapper>
+    </>
   )
 }
