+++ packages/synapse-interface/pages/state-managed-bridge/index.tsx
@@ -85,22 +85,6 @@ import {
 } from '@/slices/priceDataSlice'
 import { isTransactionReceiptError } from '@/utils/isTransactionReceiptError'
 import { SwitchButton } from '@/components/buttons/SwitchButton'
-import { useEventCountdownProgressBar } from '@/components/Maintenance/EventCountdownProgressBar'
-
-import {
-  METIS_DOWNTIME_START_DATE,
-  METIS_DOWNTIME_END_DATE,
-  MetisDowntimeWarningMessage,
-  useMetisDowntimeCountdownProgress,
-} from '@/components/Maintenance/Events/MetisUpgrade'
-import {
-  ECOTONE_FORK_START_DATE,
-  ECOTONE_FORK_END_DATE,
-  EcotoneForkWarningMessage,
-  useEcotoneForkCountdownProgress,
-} from '@/components/Maintenance/Events/EcotoneForkUpgrade'
-
-import { OPTIMISM, BASE, METIS } from '@/constants/chains/master'
 
 const StateManagedBridge = () => {
   const { address } = useAccount()
@@ -533,20 +517,6 @@ const StateManagedBridge = () => {
   const springClass =
     '-mt-4 fixed z-50 w-full h-full bg-opacity-50 bg-[#343036]'
 
-  /* Remove after upgrades */
-  const {
-    isEcotoneForkUpgradePending,
-    isCurrentChainDisabled: isEcotoneUpgradeChainsDisabled,
-    EcotoneForkCountdownProgressBar,
-  } = useEcotoneForkCountdownProgress()
-
-  const {
-    isMetisUpgradePending,
-    isCurrentChainDisabled: isMetisUpgradeChainDisabled,
-    MetisUpgradeCountdownProgressBar,
-  } = useMetisDowntimeCountdownProgress()
-  /* Remove after upgrades */
-
   return (
     <div className="flex flex-col w-full max-w-lg mx-auto lg:mx-0">
       <div className="flex flex-col">
@@ -584,8 +554,6 @@ const StateManagedBridge = () => {
             transition-all duration-100 transform rounded-md
           `}
         >
-          {EcotoneForkCountdownProgressBar}
-          {MetisUpgradeCountdownProgressBar}
           <div ref={bridgeDisplayRef}>
             <Transition show={showSettingsSlideOver} {...TRANSITION_PROPS}>
               <animated.div>
@@ -623,12 +591,6 @@ const StateManagedBridge = () => {
             />
             <OutputContainer />
             <Warning />
-
-            {/* Remove after upgrades */}
-            {isEcotoneForkUpgradePending && <EcotoneForkWarningMessage />}
-            {isMetisUpgradePending && <MetisDowntimeWarningMessage />}
-            {/* Remove after upgrades */}
-
             <Transition
               appear={true}
               unmount={false}
@@ -648,9 +610,7 @@ const StateManagedBridge = () => {
                 isApproved={isApproved}
                 approveTxn={approveTxn}
                 executeBridge={executeBridge}
-                isBridgePaused={
-                  isEcotoneUpgradeChainsDisabled || isMetisUpgradeChainDisabled
-                }
+                isBridgePaused={false}
               />
             </div>
           </div>
